refactor: 새로 타입 분리
